Adicionando trava para quando memoria disponivel for maior que a real o programa na quebre

+++ msort.h
@@ -1,17 +1,10 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <iostream>
-#include <fstream>
-#include <string>
-#include <vector>
-#include <math.h>
 #include "utils.h"
 
 using namespace std;
 
 class ExternalSorter {
 	public:
-		ExternalSorter(string &entrada, string &saida, int &memoria, int &vias ): 
+		ExternalSorter(string &entrada, string &saida, unsigned long int &memoria, int &vias ): 
 		m_arqEntrada(entrada), m_arqSaida(saida), m_memoriaDisponivel(memoria), m_qtdVias(vias), m_etapa(0)
 		{
 			// Calcula o máximo de valores que cabem na memoria
@@ -27,12 +20,12 @@ class ExternalSorter {
 		
 		string m_arqEntrada;
 		string m_arqSaida;
-		int m_memoriaDisponivel;
+		unsigned long int m_memoriaDisponivel;
 		int m_qtdVias;
 		
 		int m_totalValores;
 		// Variavel contem o máximos de elementos que cabem na memoria
-		int m_valoresMemoria;
+		unsigned long int m_valoresMemoria;
 		int m_etapa;
 		FILE* m_pFile;
 };
